--- src/components/usePageTracking.js
@@ -1,25 +1,0 @@
-import { useEffect, useState } from "react";
-import { useLocation } from "react-router-dom";
-import ReactGA from "react-ga";
-
-const usePageTracking = () => {
-  const location = useLocation();
-  const [initialized, setInitialized] = useState(false);
-
-  useEffect(() => {
-    if (!window.location.href.includes("localhost")) {
-      ReactGA.initialize('UA-183493843-1');
-    }
-    setInitialized(true);
-  }, []);
-
-  useEffect(() => {
-    if (initialized) {
-      ReactGA.pageview(location.pathname + location.search);
-    }
-  }, [initialized, location]);
-};
-
-export default usePageTracking;
-
-export default usePageTracking;
